+++ app/src/main/java/com/olshevchenko/ghjobwatcher/GitHubApiState.kt
@@ -1,0 +1,17 @@
+package com.olshevchenko.ghjobwatcher
+
+/**
+ * class displaying state of network operation with GitHub
+ */
+data class GitHubApiState private constructor(
+    val status: Status,
+    val msg: String? = null
+) {
+    companion object {
+        val RUNNING = GitHubApiState(Status.LOADING)
+        val OK = GitHubApiState(Status.DONE)
+        fun error(errMsg: String?) = GitHubApiState(Status.ERROR, msg = errMsg)
+    }
+
+    enum class Status { LOADING, ERROR, DONE }
+}