+++ app/src/main/java/com/olshevchenko/ghjobwatcher/GitHubJob.kt
@@ -1,15 +1,11 @@
 package com.olshevchenko.ghjobwatcher
 
-import android.os.Build
 import android.os.Parcelable
-import android.text.Html
-import android.text.Html.FROM_HTML_MODE_LEGACY
-import android.text.Spanned
 import android.util.Log
 import com.olshevchenko.ghjobwatcher.network.GitHubApiJob
+import com.olshevchenko.ghjobwatcher.utils.GitHubDateUtils
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import kotlinx.android.parcel.RawValue
 import java.text.ParseException
 import java.util.*
 
@@ -19,11 +15,33 @@ import java.util.*
  */
 @Parcelize
 class GitHubJob(
-    val apiJob: GitHubApiJob,
-    var createdMNDateString: String = "",
-    var howToApplySpanned: @RawValue Spanned? = null, // ToDo NOT forget to write serilizer/deserializer!!
-    var descriptionSpanned: @RawValue Spanned? = null, // ToDo ==""==
+    private val apiJob: GitHubApiJob,
+    var createdMNDateString: String = ""
 ) : Parcelable {
+
+    companion object {
+        const val EMPLOYMENT_TYPE_EMPTY = "N/A"
+        const val CREATED_AT_EMPTY = ""
+        const val COMPANY_NAME_EMPTY = "N/A"
+        const val COMPANY_URL_EMPTY = ""
+        const val JOB_TITLE_EMPTY = "N/A"
+        const val JOB_DESCRIPTION_EMPTY = "N/A"
+        const val HOW_TO_APPLY_EMPTY = "N/A"
+        const val COMPANY_LOGO_EMPTY = ""
+    }
+
+    /**
+     * Mapping from nullable API fields to clear model ones
+     */
+    val id = apiJob.id ?: throw ParseException("Job with id='null'", 0)
+    val employmentType = apiJob.type ?: EMPLOYMENT_TYPE_EMPTY
+    val createdAt = apiJob.created_at ?: CREATED_AT_EMPTY
+    val companyName = apiJob.company ?: COMPANY_NAME_EMPTY
+    val companyUrl = apiJob.company_url ?: COMPANY_URL_EMPTY
+    val jobTitle = apiJob.title ?: JOB_TITLE_EMPTY
+    val jobDescription = apiJob.description ?: JOB_DESCRIPTION_EMPTY
+    val howToApply = apiJob.how_to_apply ?: HOW_TO_APPLY_EMPTY
+    val companyLogoUrl = apiJob.company_logo ?: COMPANY_LOGO_EMPTY
 
     @IgnoredOnParcel
     private var _date: Date
@@ -31,7 +49,7 @@ class GitHubJob(
     /** parse JSON creation date if possible. Otherwise, get current date (?) */
     init {
         try {
-            _date = GitHubDateUtils.parseFromJSONDate(apiJob._createdAt)
+            _date = GitHubDateUtils.parseFromJSONDate(createdAt)
         } catch (e: ParseException) {
             Log.w(
                 "GitHubJob",
@@ -50,19 +68,16 @@ class GitHubJob(
      */
     init {
         createdMNDateString = GitHubDateUtils.formatToHomemadeDate(createdMNDate)
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            howToApplySpanned = Html.fromHtml(apiJob.howToApply, FROM_HTML_MODE_LEGACY)
-            descriptionSpanned = Html.fromHtml(apiJob.description, FROM_HTML_MODE_LEGACY)
-        } else {
-            @Suppress("DEPRECATION")
-            howToApplySpanned = Html.fromHtml(apiJob.howToApply)
-            @Suppress("DEPRECATION")
-            descriptionSpanned = Html.fromHtml(apiJob.description)
-        }
     }
 
     /** flag whether to show [createdMNDateString] of the job in UI list or not */
     @IgnoredOnParcel
     var isDateVisible = false
+//
+//    override fun compareTo(other: GitHubJob): Int {
+////        if (this.createdMNDate != other.createdMNDate) {
+////            return (this.createdMNDate - other.createdMNDate).toInt()
+////        } else return 0
+//        return 0;
+//    }
 }